refactor: model.domain 패키지를 model 패키지로 정리
