+++ src/main/java/lotto/model/WinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.model;
 
 public final class WinningNumbers {
