+++ src/main/java/lotto/model/generator/RandomNumbersGenerator.java
@@ -1,4 +1,4 @@
-package lotto.model.domain.generator;
+package lotto.model.generator;
 
 import java.util.ArrayList;
 import java.util.Collections;
