+++ src/test/java/lotto/model/LottoNumberTest.java
@@ -1,10 +1,9 @@
-package lotto.model.domain;
+package lotto.model;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
-
 
 class LottoNumberTest {
