+++ src/test/java/lotto/model/ManualLottosTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
