+++ src/test/java/lotto/model/RankTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain;
+package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
