+++ src/test/java/lotto/model/generator/RandomNumbersGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.model.domain.generator;
+package lotto.model.generator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
